requirements.txt
fastapi
uvicorn
pillow
pytesseract
opencv-python
python-multipart
